Refactor formatting of display text for TextUI
Extracted responsibility of formatting text for display to Formatter.
